fix: address code review findings for Task 1
- INSERT → INSERT IGNORE for idempotent seed data (P0)
- user_id: BIGINT UNSIGNED → INT UNSIGNED to match existing user table (P1)
- Add rollback comment block (P2)
- Add MySQL Error 1093 reference in comment (P1)

+++ internal/pkg/model/llm.go
@@ -65,7 +65,7 @@ func (LLMModelProvider) TableName() string { return "llm_model_provider" }
 // UserModelPreference 用户模型偏好
 type UserModelPreference struct {
 	ID        uint64    `gorm:"primaryKey;autoIncrement" json:"id"`
-	UserID    uint64    `gorm:"not null;uniqueIndex:uk_user_feature" json:"user_id"`
+	UserID    uint      `gorm:"not null;uniqueIndex:uk_user_feature" json:"user_id"`
 	Feature   string    `gorm:"size:20;not null;uniqueIndex:uk_user_feature" json:"feature"`
 	ModelKey  string    `gorm:"size:100;not null" json:"model_key"`
 	Thinking  bool      `gorm:"default:false" json:"thinking"`
